Move sync adapter helper classes to their own files so things stay tidy
Need to get some units for each of them as well.
